Fixed a couple of bugs and added some of the monkeys

+++ Assets/Scripts/Simulation/UI/MoneyController.cs
@@ -22,7 +22,7 @@ public class MoneyController : MonoBehaviour {
 	private void UpdateText() {
 		string spent = Format(Simulation.currentSave.spent);
 		string target = Format(Simulation.currentSave.spendTarget);
-		int percent = Mathf.FloorToInt((Simulation.currentSave.spent / Simulation.currentSave.spendTarget) * 100);
+		int percent = Mathf.FloorToInt(((float)Simulation.currentSave.spent / Simulation.currentSave.spendTarget) * 100);
 
 		text.SetText($"{spent}/£{target} ({percent}%)");
 	}
